+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - bug fixes
   - meter's properties not accessible (overridden with fader's properties)
+  - `--read-only` option not properly working
 - mobile
   - button for preventing device sleep